chore: Move `Switch` component
- This doesn't really need a redesign.
- We did try using the Reanimated's new CSS transitions, but we did see some jitters.
  - On the Sort Sheet, the animation just blinks to the end because the JS thread is blocked from re-ordering the list contents.
